refactor(acp-client): extract exception hierarchy into errors.py
client.py was 937 lines; the top ~170 were the exception classes
(AcpErrorCode, AcpClientError, AcpServerNotFound, AcpRpcError,
AcpInteractionRequiredError), orthogonal to the JSON-RPC transport
machinery. Move them to acp_client/errors.py and re-import them into
client.py.

The names are re-imported into client.py's namespace (and stay exported
from the package __init__), so every existing call site keeps working:
`from agentao.acp_client.client import AcpClientError`, the manager
submodules' `from ..client import AcpServerNotFound`, and
`from agentao.acp_client import AcpRpcError` all resolve to the same
class objects (verified: identity + subclass hierarchy intact). Classes
moved verbatim (semantic diff shows zero logic change). Dropped the
now-unused `from enum import Enum` from client.py. client.py: 937 → 779
lines. Full suite green (2840 passed).
